backend/requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
Django==4.1.7
djangorestframework==3.14.0
flake8==6.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
psycopg2-binary
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3
sqlparse==0.4.3
RateMyProfessorAPI==1.3.6